Add responsive styles for Login, register & session components
