--- src/components/Session/session.css
@@ -1,11 +1,0 @@
-.auth-box {
-  display: flex;
-  width: 100%;
-  justify-content: center;
-  align-items: center;
-  height: 100vh;
-}
-
-.row-box {
-  width: 50%;
-}
